Feature/permissions

+++ src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt =  require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../../src/mail');
+const { hasPermission }  = require('../utils');
 
 const generateJWT = function(userID) {
     return jwt.sign({userId: userID }, process.env.APP_SECRET);
@@ -41,11 +42,18 @@ const mutations = {
     },
 
     async deleteItem(parent, args, ctx, info) {
-        const where = {id : args.id};
         //1. find the item
-        const item = await ctx.db.query.item({ where }, `{id title}`);
+        const where = { id : args.id };
+        const item = await ctx.db.query.item({ where }, `{id title user {id} }`);
+    
         //2. Check if they own the item
-        //TODO
+        const ownsItem = item.user.id === ctx.request.userID;
+        const hasPermission = ctx.request.user.permissions.some(permission => ["ITEMDELETE", "ADMIN"].includes(permission));
+
+        if(!ownsItem && !hasPermission) {
+            throw new Error("You don't have permission to do that");
+        }
+
         //3. Delete it
         return ctx.db.mutation.deleteItem({ where }, info);
     },
@@ -201,7 +209,37 @@ const mutations = {
         });
         // 8. return new user
         return updatedUser;
-    }
+    },
+
+
+    async updatePermissions(parent, args, ctx, info) {
+        //check if they're logged in
+        if(!ctx.request.userID) {
+            throw new Error("You need to be logged in to do that");
+        }
+        //query the current user
+        const user = await ctx.db.query.user({
+            where: {
+            id:ctx.request.userID
+        }}, info);
+
+        //check if they're allowed to do this
+        hasPermission(user, ["ADMIN", "PERMISSIONUPDATE"]);
+
+        //update permissions
+        console.log(args);
+        return ctx.db.mutation.updateUser({
+            data: {
+                permissions: {
+                    set: args.permissions
+                }
+            },
+            where: {
+                id: args.where.id
+            }
+        }, info);
+
+    },
 
 };
